Also show media message if trying to play modern video on a legacy MAP
